Fix rebuilding of timeline dropdowns to show full reports timeframe
Timeline dropdowns should show the full span from earliest to latest report
regardless of interval (months/day). This fixes an issue where timeline
was limited to the same year as the active display.

Also fixes the selected dates to be the active display dates, not the
start and end of the timeline.